Add root verifier in general unit tests
